+++ SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialActorChannel.cpp
@@ -256,7 +256,7 @@ void USpatialActorChannel::RetireEntityIfAuthoritative()
 			}
 			else
 			{
-				Sender->RetireEntity(EntityId, Actor->IsNetStartupActor());
+				NetDriver->ActorSystem->RetireEntity(EntityId, Actor->IsNetStartupActor());
 			}
 		}
 		else if (bCreatedEntity) // We have not gained authority yet
@@ -565,6 +565,9 @@ int64 USpatialActorChannel::ReplicateActor()
 	UE_LOG(LogNetTraffic, Log, TEXT("Replicate %s, bNetInitial: %d, bNetOwner: %d"), *Actor->GetName(), RepFlags.bNetInitial,
 		   RepFlags.bNetOwner);
 
+	// Always replicate initial only properties and rely on QBI to filter where necessary.
+	RepFlags.bNetInitial = true;
+
 	FMemMark MemMark(FMemStack::Get()); // The calls to ReplicateProperties will allocate memory on FMemStack::Get(), and use it in
 										// ::PostSendBunch. we free it below
 
@@ -583,7 +586,7 @@ int64 USpatialActorChannel::ReplicateActor()
 	{
 		if (SpatialGDKSettings->bBatchSpatialPositionUpdates)
 		{
-			Sender->RegisterChannelForPositionUpdate(this);
+			NetDriver->ActorSystem->RegisterChannelForPositionUpdate(this);
 		}
 		else
 		{
@@ -658,7 +661,7 @@ int64 USpatialActorChannel::ReplicateActor()
 
 	if (!bCreatingNewEntity && NeedOwnerInterestUpdate() && NetDriver->InterestFactory->DoOwnersHaveEntityId(Actor))
 	{
-		Sender->UpdateInterestComponent(Actor);
+		NetDriver->ActorSystem->UpdateInterestComponent(Actor);
 		SetNeedOwnerInterestUpdate(false);
 	}
 
@@ -671,7 +674,7 @@ int64 USpatialActorChannel::ReplicateActor()
 			// so we know what subobjects are relevant for replication when creating the entity.
 			Actor->ReplicateSubobjects(this, &Bunch, &RepFlags);
 
-			Sender->SendCreateEntityRequest(this, ReplicationBytesWritten);
+			NetDriver->ActorSystem->SendCreateEntityRequest(this, ReplicationBytesWritten);
 
 			bCreatedEntity = true;
 
@@ -687,7 +690,7 @@ int64 USpatialActorChannel::ReplicateActor()
 		{
 			FRepChangeState RepChangeState = { RepChanged, GetObjectRepLayout(Actor) };
 
-			Sender->SendComponentUpdates(Actor, Info, this, &RepChangeState, &HandoverChangeState, ReplicationBytesWritten);
+			NetDriver->ActorSystem->SendComponentUpdates(Actor, Info, this, &RepChangeState, &HandoverChangeState, ReplicationBytesWritten);
 
 			bInterestDirty = false;
 		}
@@ -749,8 +752,8 @@ int64 USpatialActorChannel::ReplicateActor()
 			FHandoverChangeState SubobjectHandoverChangeState = GetHandoverChangeList(SubobjectHandoverShadowData->Get(), Subobject);
 			if (SubobjectHandoverChangeState.Num() > 0)
 			{
-				Sender->SendComponentUpdates(Subobject, SubobjectInfo, this, nullptr, &SubobjectHandoverChangeState,
-											 ReplicationBytesWritten);
+				NetDriver->ActorSystem->SendComponentUpdates(Subobject, SubobjectInfo, this, nullptr, &SubobjectHandoverChangeState,
+															 ReplicationBytesWritten);
 			}
 		}
 
@@ -765,8 +768,8 @@ int64 USpatialActorChannel::ReplicateActor()
 				{
 					OnSubobjectDeleted(ObjectRef, RepComp.Key(), RepComp.Value()->GetWeakObjectPtr());
 
-					Sender->SendRemoveComponentForClassInfo(EntityId,
-															NetDriver->ClassInfoManager->GetClassInfoByComponentId(ObjectRef.Offset));
+					NetDriver->ActorSystem->SendRemoveComponentForClassInfo(
+						EntityId, NetDriver->ClassInfoManager->GetClassInfoByComponentId(ObjectRef.Offset));
 				}
 
 				RepComp.Value()->CleanUp();
@@ -822,7 +825,7 @@ void USpatialActorChannel::DynamicallyAttachSubobject(UObject* Object)
 
 	check(Info != nullptr);
 
-	Sender->SendAddComponentForSubobject(this, Object, *Info, ReplicationBytesWritten);
+	NetDriver->ActorSystem->SendAddComponentForSubobject(this, Object, *Info, ReplicationBytesWritten);
 }
 
 bool USpatialActorChannel::ReplicateSubobject(UObject* Object, const FReplicationFlags& RepFlags)
@@ -925,7 +928,7 @@ bool USpatialActorChannel::ReplicateSubobject(UObject* Object, const FReplicatio
 		}
 
 		const FClassInfo& Info = NetDriver->ClassInfoManager->GetOrCreateClassInfoByObject(Object);
-		Sender->SendComponentUpdates(Object, Info, this, &RepChangeState, nullptr, ReplicationBytesWritten);
+		NetDriver->ActorSystem->SendComponentUpdates(Object, Info, this, &RepChangeState, nullptr, ReplicationBytesWritten);
 
 		SendingRepState->HistoryEnd++;
 	}
@@ -1041,6 +1044,7 @@ FHandoverChangeState USpatialActorChannel::GetHandoverChangeList(TArray<uint8>& 
 void USpatialActorChannel::SetChannelActor(AActor* InActor, ESetChannelActorFlags Flags)
 {
 	Super::SetChannelActor(InActor, Flags);
+	check(NetDriver->GetSpatialOSNetConnection() == Connection);
 	USpatialPackageMapClient* PackageMap = NetDriver->PackageMap;
 	EntityId = PackageMap->GetEntityIdFromObject(InActor);
 
@@ -1192,7 +1196,7 @@ void USpatialActorChannel::OnCreateEntityResponse(const Worker_CreateEntityRespo
 
 			// TODO: UNR-664 - Track these bytes written to use in saturation.
 			uint32 BytesWritten = 0;
-			Sender->SendCreateEntityRequest(this, BytesWritten);
+			NetDriver->ActorSystem->SendCreateEntityRequest(this, BytesWritten);
 		}
 		break;
 	case WORKER_STATUS_CODE_APPLICATION_ERROR:
@@ -1280,7 +1284,8 @@ void USpatialActorChannel::SendPositionUpdate(AActor* InActor, Worker_EntityId I
 {
 	if (InEntityId != SpatialConstants::INVALID_ENTITY_ID && NetDriver->HasServerAuthority(InEntityId))
 	{
-		Sender->SendPositionUpdate(InEntityId, NewPosition);
+		FWorkerComponentUpdate Update = SpatialGDK::Position::CreatePositionUpdate(SpatialGDK::Coordinates::FromFVector(NewPosition));
+		NetDriver->Connection->SendComponentUpdate(EntityId, &Update);
 	}
 
 	for (const auto& Child : InActor->Children)
@@ -1359,14 +1364,14 @@ void USpatialActorChannel::ServerProcessOwnershipChange()
 	}
 
 	// Owner changed, update the actor's interest over it.
-	Sender->UpdateInterestComponent(Actor);
+	NetDriver->ActorSystem->UpdateInterestComponent(Actor);
 	SetNeedOwnerInterestUpdate(!NetDriver->InterestFactory->DoOwnersHaveEntityId(Actor));
 
 	// Changing owner can affect which interest bucket the Actor should be in so we need to update it.
 	const Worker_ComponentId NewInterestBucketComponentId = NetDriver->ClassInfoManager->ComputeActorInterestComponentId(Actor);
 	if (SavedInterestBucketComponentID != NewInterestBucketComponentId)
 	{
-		Sender->SendInterestBucketComponentChange(EntityId, SavedInterestBucketComponentID, NewInterestBucketComponentId);
+		NetDriver->ActorSystem->SendInterestBucketComponentChange(EntityId, SavedInterestBucketComponentID, NewInterestBucketComponentId);
 		SavedInterestBucketComponentID = NewInterestBucketComponentId;
 		bUpdatedThisActor = true;
 	}