+++ SpatialGDK/Source/SpatialGDKEditor/Private/SchemaGenerator/SchemaGenerator.cpp
@@ -31,6 +31,10 @@ ESchemaComponentType PropertyGroupToSchemaComponentType(EReplicatedPropertyGroup
 	{
 		return SCHEMA_OwnerOnly;
 	}
+	else if (Group == REP_InitialOnly)
+	{
+		return SCHEMA_InitialOnly;
+	}
 	else
 	{
 		checkNoEntry();
@@ -145,8 +149,8 @@ FActorSpecificSubobjectSchemaData GenerateSchemaForStaticallyAttachedSubobject(F
 	{
 		// Since it is possible to replicate subobjects which have no replicated properties.
 		// We need to generate a schema component for every subobject. So if we have no replicated
-		// properties, we only don't generate a schema component if we are REP_SingleClient
-		if (RepData[Group].Num() == 0 && Group == REP_SingleClient)
+		// properties, we only generate a schema component if we are REP_MultiClient.
+		if (RepData[Group].Num() == 0 && Group != REP_MultiClient)
 		{
 			continue;
 		}
@@ -429,8 +433,8 @@ void GenerateSubobjectSchema(FComponentIdGenerator& IdGenerator, UClass* Class, 
 	{
 		// Since it is possible to replicate subobjects which have no replicated properties.
 		// We need to generate a schema component for every subobject. So if we have no replicated
-		// properties, we only don't generate a schema component if we are REP_SingleClient
-		if (RepData[Group].Num() == 0 && Group == REP_SingleClient)
+		// properties, we only generate a schema component if we are REP_MultiClient.
+		if (RepData[Group].Num() == 0 && Group != REP_MultiClient)
 		{
 			continue;
 		}
@@ -502,8 +506,8 @@ void GenerateSubobjectSchema(FComponentIdGenerator& IdGenerator, UClass* Class, 
 		{
 			// Since it is possible to replicate subobjects which have no replicated properties.
 			// We need to generate a schema component for every subobject. So if we have no replicated
-			// properties, we only don't generate a schema component if we are REP_SingleClient
-			if (RepData[Group].Num() == 0 && Group == REP_SingleClient)
+			// properties, we only generate a schema component if we are REP_MultiClient.
+			if (RepData[Group].Num() == 0 && Group != REP_MultiClient)
 			{
 				continue;
 			}