+++ SpatialGDK/Source/SpatialGDKFunctionalTests/Private/TestMaps/GeneratedTestMap.cpp
@@ -1,12 +1,14 @@
 // Copyright (c) Improbable Worlds Ltd, All Rights Reserved
 
 #include "TestMaps/GeneratedTestMap.h"
+#include "Core/Public/Misc/FileHelper.h"
 #include "Engine/ExponentialHeightFog.h"
 #include "Engine/SkyLight.h"
 #include "Engine/StaticMeshActor.h"
 #include "EngineClasses/SpatialWorldSettings.h"
 #include "FileHelpers.h"
 #include "GameFramework/PlayerStart.h"
+#include "SpatialGDKEditor/Public/SpatialTestSettings.h"
 #include "Tests/AutomationEditorCommon.h"
 #include "UObject/ConstructorHelpers.h"
 
@@ -35,13 +37,26 @@ AActor* UGeneratedTestMap::AddActorToLevel(ULevel* Level, UClass* Class, const F
 	return GEditor->AddActor(Level, Class, Transform);
 }
 
-void UGeneratedTestMap::GenerateMap()
+bool UGeneratedTestMap::GenerateMap()
 {
 	checkf(bIsValidForGeneration, TEXT("This test map object is not valid for map generation, please use the UGeneratedTestMap constructor "
 									   "with arguments when deriving from the base UGeneratedTestMap."));
 	GenerateBaseMap();
 	CreateCustomContentForMap();
-	SaveMap();
+	return SaveMap();
+}
+
+bool UGeneratedTestMap::GenerateCustomConfig()
+{
+	if (CustomConfigString.IsEmpty())
+	{
+		// Only create a custom config file if we have something meaningful to write so we don't pollute the file system too much
+		return true;
+	}
+
+	const FString OverrideSettingsFilename =
+		FSpatialTestSettings::GeneratedOverrideSettingsBaseFilename + MapName + FSpatialTestSettings::OverrideSettingsFileExtension;
+	return FFileHelper::SaveStringToFile(CustomConfigString, *OverrideSettingsFilename);
 }
 
 FString UGeneratedTestMap::GetGeneratedMapFolder()
@@ -70,19 +85,15 @@ void UGeneratedTestMap::GenerateBaseMap()
 
 	APlayerStart* PlayerStart = AddActorToLevel<APlayerStart>(CurrentLevel, FTransform(FVector(-500, 0, 100)));
 
-	// Use current settings for all these generated maps as default.
-	// This will probably change with UNR-4801.
-	ASpatialWorldSettings* WorldSettings = CastChecked<ASpatialWorldSettings>(World->GetWorldSettings());
-	WorldSettings->TestingSettings.TestingMode = EMapTestingMode::UseCurrentSettings;
-
 	// TODO: Maybe figure out how to set the default viewpoint when opening the map, so we don't start in the ground (maybe together with
 	// other viewing position issues), ticket: UNR-4975.
 }
 
-void UGeneratedTestMap::SaveMap()
+bool UGeneratedTestMap::SaveMap()
 {
 	const bool bSuccess = FEditorFileUtils::SaveLevel(World->GetCurrentLevel(), GetPathToSaveTheMap());
 	UE_CLOG(!bSuccess, LogGenerateTestMapsCommandlet, Error, TEXT("Failed to save the map %s."), *GetMapName());
+	return bSuccess;
 }
 
 FString UGeneratedTestMap::GetPathToSaveTheMap()