+++ SpatialGDK/Source/SpatialGDKFunctionalTests/SpatialGDK/SpatialTestSingleServerDynamicComponents/SpatialTestSingleServerDynamicComponents.h
@@ -27,4 +27,5 @@ public:
 	UTestDynamicComponent* CreateAndAttachTestDynamicComponentToActor(AActor* Actor, FName Name);
 
 	const FVector ActorSpawnPosition = FVector(0.0f, 0.0f, 50.0f);
+	bool bInitialOnlyEnabled;
 };